feat: create main page frame
