+++ lib/widgets/pages/main_page.dart
@@ -1,0 +1,40 @@
+import 'package:flutter/material.dart';
+
+
+class MainPage extends StatefulWidget {
+  const MainPage({Key? key}) : super(key: key);
+
+  @override
+  State<MainPage> createState() => _MainPage();
+}
+
+class _MainPage extends State<MainPage> {
+  int _counter = 0;
+
+  void _incrementCounter() {
+    setState(() {
+      _counter++;
+    });
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: PreferredSize(
+        preferredSize: const Size.fromHeight(100.0),
+        child: AppBar(
+          title: const Image(
+            width: 210,
+            image: AssetImage('images/tiny_news.png'),
+            fit: BoxFit.contain
+          )
+          ),
+        ),
+      body: Column(
+        children: [
+
+        ],
+      )
+    );
+  }
+}
